workspace/pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["J1ng89Styl3 <jingyue8999xe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.5.0"
numpy = "^1.26.0"
flask = "^3.0.3"
flask-wtf = "^1.2.1"
jinja2 = "^3.1.4"
flask-login = "^0.6.3"
flask-httpauth = "^4.8.0"
flask-security = "^5.5.2"
flask-restful = "^0.3.10"
marshmallow = "^3.22.0"
celery = "^5.4.0"
redis = "^5.0.8"
flake8 = "^7.1.1"
black = "^24.8.0"
mypy = "^1.11.2"
isort = "^5.13.2"
flask-sqlalchemy = "^3.1.1"
openai = "^1.40.0"
load-dotenv = "^0.1.0"
datetime = "^5.5"
uuid7 = "^0.1.0"
graphene = "^3.3"
graphene-sqlalchemy = "3.0.0rc1"
pandas = "^2.2.2"
flask-cors = "^4.0.1"
langchain = "^0.2.15"
langchain-core = "^0.2.36"
azure-identity = "^1.17.1"
langchain-community = "^0.2.15"
langchain-openai = "^0.1.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
